Add accuracy to lat lon data #371
add accuracy field to location, all report gps (case, contact, event, visit) and "task done"
add report lat lon to sample
show accuracy for locations

+++ sormas-api/src/main/java/de/symeda/sormas/api/facility/FacilityDto.java
@@ -23,8 +23,8 @@ public class FacilityDto extends FacilityReferenceDto {
 	private DistrictReferenceDto district;
 	private CommunityReferenceDto community;
 	private String city;
-	private Float latitude;
-	private Float longitude;
+	private Double latitude;
+	private Double longitude;
 	private FacilityType type;
 	private boolean publicOwnership;
 	
@@ -58,17 +58,17 @@ public class FacilityDto extends FacilityReferenceDto {
 		this.community = community;
 	}
 	
-	public Float getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(Float latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 	
-	public Float getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
-	public void setLongitude(Float longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
